py/kernel/cloud/scripts/requirements.txt
google-cloud-batch
google-cloud-storage
google-cloud-functions
functions-framework